feat(build): Including the falco-driver-loader into the container image
So I might be wrong here, but I believe this is missing.

I tried to build to the `docker/falco/Dockerfile` locally and push to GKE and was
running into some errors getting Falco started.

I checked both the GKE node, as well as the running pod and I couldn't find the script.

So I think we want to include the script directly into the container.

Anyway I was able to now use this Dockerfile to run my local version of Falco in GKE and
load a BPF probe fine.

Other thoughts: Do we want to consider pull the flags from the `falco-driver-loader` script
up to the container image as an ENV or something?

Other thoughts: It looks like the other container images all are based on this one so this
should naturally flow down to the others.

If we don't need this feel free to ignore/close. :)

Signed-off-by: Kris Nova <kris@nivenly.com>