Initial skeleton of "action" object.
An "action" is a piece of code that has a name (e.g. "load plugins")
and can name dependencies (depends on "init inspector").

A run() method is called to do the work of the action, and a deinit()
method is called after all actions are complete to tear down any state
outside the object.

actions can be segregated into groups (e.g. "init" vs "run"). All
actions in a given group are run before actions in the next group. The
action manager is configured with the group order.

Actions have access to to command line options via options() and can
share state with other actions via state().

This will help distribute all the many things falco does in falco_init
into smaller components, while retaining a notion of dependencies and
order.

To make testing easier, most of the functionality is in a base class
::runnable_action, which defines the interface. A derived class
::action additionally brings in an application object and the
state()/options() methods.

This makes it easier to write unit tests for the action manager
without bringing in all of application, falco engine, outputs,
inspector, etc.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>