Add direct access to source, without copying
The current methods find_source() that return a reference to a
falco_source perform a copy, as they dereference the pointer returned
by .at().

In some cases, the caller wants to directly modify the value inside
m_sources, so create a version that returns a direct pointer instead
of a reference.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>