Set Default Driver Download Location
TL;DR: Set a default driver download location to
https://download.falco.org/driver to facilitate a more
download-and-go user experience for the driver-loader script.

Every few months or so I come back to the falco project and need
to get reacquainted with the structure of where things are. How do
I install the kernel modules? Where should I download the kernel
modules from? Where are the eBPF probes? It's pretty consistently
20 minutes to an hour of work that needs to be repeated two or
more times per year.

One of the things that promised to help with this was the
falco-driver-loader script. It's really good! As soon as you know
all the inputs to the script. So instead of relying on future me
to reverse engineer the inputs from a lack of knowledge, present
me is going to update the script to include the public
distribution point of eBPF probes and kernel modules.

Signed-off-by: Tom Kelley <distortedsignal@gmail.com>